shopping cart added & tested
a shopping cart where you can add products which then reflect in the checkout. If you add more quantity then it should increase the quantity of the same product. If you add another product then it should show the product. You can use the static images or items as products.